Remove symbol dependencies from process object
