+++ include/dwfl_hdr.hpp
@@ -60,17 +60,4 @@ struct DwflWrapper {
   std::unordered_map<FileInfoId_t, DDProfMod> _ddprof_mods;
 };
 
-class DwflHdr {
-public:
-  // checks against maximum number of PIDs
-  void clear_pid(pid_t pid);
-
-  // get number of accessed modules
-  int get_nb_mod() const;
-  void display_stats() const;
-
-private:
-  std::unordered_map<pid_t, DwflWrapper> _dwfl_map;
-};
-
 } // namespace ddprof
