+++ include/unwind_helper.hpp
@@ -29,9 +29,6 @@ void add_dso_frame(UnwindState *us, const Dso &dso,
 
 void add_virtual_base_frame(UnwindState *us);
 
-bool memory_read(ProcessAddress_t addr, ElfWord_t *result, int regno,
-                 void *arg);
-
 void add_error_frame(const Dso *dso, UnwindState *us, ProcessAddress_t pc,
                      SymbolErrors error_case = SymbolErrors::unknown_dso);
