+++ src/ddprof_worker.cc
@@ -21,7 +21,7 @@
 #include "tags.hpp"
 #include "tsc_clock.hpp"
 #include "unwind.hpp"
-#include "unwind_helpers.hpp"
+#include "unwind_helper.hpp"
 #include "unwind_state.hpp"
 
 #include <algorithm>
